feature-store/feast_server_deploy/kafka_demo/requirements.txt
kafka-python==2.0.2
pandas
pyarrow
